AutoComplete
AutoComplete  text view that shows completion suggestions automatically
while the user is typing. The list of suggestions is displayed in a drop
down menu from which the user can choose an item to replace the content
of the edit box with.

The drop down can be dismissed at any time by pressing the back key or,
if no item is selected in the drop down, by pressing the enter/dpad
center key.

The list of suggestions is obtained from a data adapter and appears only
after a given number of characters defined by the threshold.

I would personaly suggest to use MultiAutoCompleteTextView than
AutoComplete because it can
show completion suggestions for the substring of the text where the user
is typing instead of necessarily for the entire thing.